refactor: rename get-skills DTO and fix validation message
- Rename src/skills/dto/get-skills.ts to get-skills.dto.ts to follow project naming conventions.
- Update import references in SkillsController and SkillsService.
- Fix a typo in the Max validation decorator error message for the limit field.
